+++ src/lib/types.ts
@@ -1,0 +1,7 @@
+interface GridObject {
+    type: string;
+    unlockDate: Date;
+    pomoDoroTime: number;
+    x: number;
+    z: number;
+}