+++ Changelog.md
@@ -2,6 +2,21 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.2.0] - 2025-11-15
+
+### Added
+
+- Added Lighting
+- Upgraded timer input
+- After each pomodoro session a grass block is added
+
+## [0.1.0] - 2025-11-14
+
+### Added
+
+- Changed menu camera position
+- Added grid system for blocks
+
 ## [0.0.1] - 2025-11-14
 
 ### Added